requirements.txt
google-cloud-secret-manager

